Fix error display, use resource Type in messages
